+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Fixed
 
 - `SimpleStreams.SimpleReaderBase.readinto()` - fixed plugin invocation bug
+- added missing `show_log` argument to `image.filter()`
+
 ## [0.7.0] - 2022-08-24
 
 ### Added